LayoutTests/imported/w3c/web-platform-tests/XMLHttpRequest/formdata-foreach-expected.txt

CONSOLE MESSAGE: line 16: TypeError: fd.delete is not a function. (In 'fd.delete('n2')', 'fd.delete' is undefined)

Harness Error (FAIL), message = TypeError: fd.delete is not a function. (In 'fd.delete('n2')', 'fd.delete' is undefined)
